+++ src/hooks/useGamepadInfo.ts
@@ -1,27 +1,29 @@
 import { useState, useEffect } from 'react';
 import { transformGamepadData } from '../utils/appBusinessLogic';
+import { useGamepadPlugin } from './useGamepadPlugin';
+import type { GamepadInfo } from '../types/gamepad';
 
 export const useGamepadInfo = () => {
-  const [gamepads, setGamepads] = useState<Array<{ index: number; id: string }>>([]);
-  
+  const [gamepads, setGamepads] = useState<GamepadInfo[]>([]);
+  const { getGamepads } = useGamepadPlugin();
+
   useEffect(() => {
-    const checkGamepads = async () => {
-      if (typeof window !== 'undefined' && '__TAURI_INTERNALS__' in window) {
-        try {
-          const gamepadModule = await import('tauri-plugin-gamepad-api');
-          const currentGamepads = [...gamepadModule.getGamepads()];
-          const gamepadInfos = transformGamepadData(currentGamepads);
-          setGamepads(gamepadInfos);
-        } catch (err) {
-          console.error('Failed to get gamepads:', err);
-        }
+    if (!getGamepads) return;
+
+    const checkGamepads = () => {
+      try {
+        const currentGamepads = [...getGamepads()];
+        const gamepadInfos = transformGamepadData(currentGamepads);
+        setGamepads(gamepadInfos);
+      } catch (err) {
+        console.error('Failed to get gamepads:', err);
       }
     };
-    
+
     checkGamepads();
     const interval = setInterval(checkGamepads, 500);
     return () => clearInterval(interval);
-  }, []);
-  
+  }, [getGamepads]);
+
   return gamepads;
 };
