+++ src/utils/appBusinessLogic.test.ts
@@ -1,110 +1,11 @@
 import { describe, it, expect } from 'vitest';
 import {
-  determineDisplayStatus,
   shouldStartAutoAssignment,
   shouldConnectWebSocketForDP,
-  determineModeChangeActions,
-  playerSideToString,
-  stringToPlayerSide,
-  transparencyToString,
-  stringToTransparency,
   transformGamepadData,
   buildDPAssignments,
-  type GamepadInfo,
 } from './appBusinessLogic';
-import type { ControllerStatus, DPControllerStatus } from '../types/controller';
-
-// モックデータ作成ヘルパー
-const createMockControllerStatus = (): ControllerStatus => ({
-  keys: Array(7).fill({ isPressed: false, strokeCount: 0 }),
-  scratch: { 
-    state: 0, 
-    count: 0,
-    currentAxes: 0,
-    previousAxes: 0,
-    fixedStateTime: 0,
-    rotationDistance: 0,
-    rotationTime: 0,
-    strokeDistance: 0
-  },
-  record: { 
-    releaseTimes: [], 
-    pressedTimes: [],
-    keyReleaseTimes: [[], [], [], [], [], [], []],
-    scratchTimes: [],
-    scratchRotationDistances: []
-  }
-});
-
-const createMockDPControllerStatus = (): DPControllerStatus => ({
-  mode: 'DP',
-  player1: createMockControllerStatus(),
-  player2: createMockControllerStatus(),
-  timestamp: Date.now()
-});
-
-describe('determineDisplayStatus', () => {
-  const mockSPStatus = createMockControllerStatus();
-  const mockDPStatus = createMockDPControllerStatus();
-
-  it('受信モードでSPデータを受信した場合、そのデータを返す', () => {
-    const receivedData = createMockControllerStatus();
-    const result = determineDisplayStatus(
-      true, // isReceiveMode
-      receivedData,
-      'SP',
-      mockSPStatus,
-      mockDPStatus
-    );
-    expect(result).toBe(receivedData);
-  });
-
-  it('受信モードでDPデータを受信した場合、DPControllerStatusとして返す', () => {
-    const receivedData = createMockDPControllerStatus();
-    const result = determineDisplayStatus(
-      true, // isReceiveMode
-      receivedData,
-      'SP',
-      mockSPStatus,
-      mockDPStatus
-    );
-    expect(result).toBe(receivedData);
-    expect(result).toHaveProperty('mode', 'DP');
-  });
-
-  it('受信モードでない場合、SPモードならspControllerStatusを返す', () => {
-    const result = determineDisplayStatus(
-      false, // isReceiveMode
-      null,
-      'SP',
-      mockSPStatus,
-      mockDPStatus
-    );
-    expect(result).toBe(mockSPStatus);
-  });
-
-  it('受信モードでない場合、DPモードならdpControllerStatusを返す', () => {
-    const result = determineDisplayStatus(
-      false, // isReceiveMode
-      null,
-      'DP',
-      mockSPStatus,
-      mockDPStatus
-    );
-    expect(result).toBe(mockDPStatus);
-  });
-
-  it('受信モードだがデータがない場合、ローカルステータスを返す', () => {
-    const result = determineDisplayStatus(
-      true, // isReceiveMode
-      null, // receivedData
-      'SP',
-      mockSPStatus,
-      mockDPStatus
-    );
-    expect(result).toBe(mockSPStatus);
-  });
-});
+import type { GamepadInfo } from '../types/gamepad';
 
 describe('shouldStartAutoAssignment', () => {
   it('DPモードで両方のゲームパッドが未割り当ての場合、trueを返す', () => {
@@ -228,80 +129,6 @@ describe('shouldConnectWebSocketForDP', () => {
       false  // wsConnected
     );
     expect(result).toBe(false);
-  });
-});
-
-describe('determineModeChangeActions', () => {
-  it('同じモードの場合、何もアクションを起こさない', () => {
-    const result = determineModeChangeActions('SP', 'SP', false);
-    expect(result).toEqual({
-      shouldCancelAutoAssignment: false,
-      shouldResetCount: false,
-      shouldResetGamepad: false,
-      shouldResetDPMapping: false,
-      shouldStartAutoAssignment: false,
-    });
-  });
-
-  it('DPからSPへの切り替え時、適切なアクションを返す', () => {
-    const result = determineModeChangeActions('DP', 'SP', false);
-    expect(result).toEqual({
-      shouldCancelAutoAssignment: false,
-      shouldResetCount: true,
-      shouldResetGamepad: true,
-      shouldResetDPMapping: false,
-      shouldStartAutoAssignment: false,
-    });
-  });
-
-  it('DPからSPへの切り替え時、自動割り当て中ならキャンセルする', () => {
-    const result = determineModeChangeActions('DP', 'SP', true);
-    expect(result).toEqual({
-      shouldCancelAutoAssignment: true,
-      shouldResetCount: true,
-      shouldResetGamepad: true,
-      shouldResetDPMapping: false,
-      shouldStartAutoAssignment: false,
-    });
-  });
-
-  it('SPからDPへの切り替え時、適切なアクションを返す', () => {
-    const result = determineModeChangeActions('SP', 'DP', false);
-    expect(result).toEqual({
-      shouldCancelAutoAssignment: false,
-      shouldResetCount: true,
-      shouldResetGamepad: true,
-      shouldResetDPMapping: true,
-      shouldStartAutoAssignment: true,
-    });
-  });
-});
-
-describe('プレイヤーサイドの変換', () => {
-  it('playerSideToString が正しく変換する', () => {
-    expect(playerSideToString(false)).toBe('1P');
-    expect(playerSideToString(true)).toBe('2P');
-  });
-
-  it('stringToPlayerSide が正しく変換する', () => {
-    expect(stringToPlayerSide('1P')).toBe(false);
-    expect(stringToPlayerSide('2P')).toBe(true);
-    expect(stringToPlayerSide(null)).toBe(false);
-    expect(stringToPlayerSide('invalid')).toBe(false);
-  });
-});
-
-describe('透過状態の変換', () => {
-  it('transparencyToString が正しく変換する', () => {
-    expect(transparencyToString(true)).toBe('true');
-    expect(transparencyToString(false)).toBe('false');
-  });
-
-  it('stringToTransparency が正しく変換する', () => {
-    expect(stringToTransparency('true')).toBe(true);
-    expect(stringToTransparency('false')).toBe(false);
-    expect(stringToTransparency(null)).toBe(true); // デフォルトは透過
-    expect(stringToTransparency('invalid')).toBe(true);
   });
 });
